Conflict resolved and refactor
